Add GCF VM Delete Event Handler Example
Customer use case: When a VM is deleted, DNS A records associated with the VM
should be automatically deleted so that stale DNS records do not pile up.

The customer requested the solution be general enough to support VMs deleted
through the Console UI, gcloud, Deployment Manager, Terraform, etc...

The solution uses a GCF Background Function handling `compute.instances.delete
events`.  The customer also required the function compare the IP address of the
VM's primary interface with the DNS A record rrdatas to guard against deleting
the incorrect record.  This IP check introduces a race condition, the function
races against the VM delete operation.  In practice the function wins the race
most of the time and the cleanup is successful.