Improve log UX when DNS cleanup fails for reason IP_NOT_AVAILABLE
This patch is intended to enable the customer to easily identify when
the DNS A record cleanup is not performed because the IP address cannot
be determined.  The goal is to provide structured data to easily filter
and identify VM instances which did not have their A record deleted.

The reason: 'IP_NOT_AVAILABLE' key/value is intended to provide this
filter.