Refactor to DnsVmGc app class and structured logs
App class
---

This patch refactors GCF function into a DnsVmGc class.  Instances of
this class are intended to have the same lifetime as the GCF background
function process.  API discovery happens once during cold start and is
re-used across multiple trigger events.

The EventHandler class models a single event processing execution.
Helper methods ensure results are reported consistently via the
`log_struct` method.  Result logs are logged into the
`reports/dns_vm_gc` stream.

Structured Logs
---

Without this patch the log structure data is sprinkled throughout the
code.  This is a problem because it's tangled up with the behavior,
instead the structured logs should be considered a contract with the end
user who will generate reports based on the logs.

This patch defines a set of StructuredLog sub-classes modeling the
structured log data.  This separates the log structure from the behavior
of the code, making it easier to change the behavior of the code without
affecting the log structure the user relies on.

Structured Logs are sent to the
`projects/$FUNCTION_PROJECT/logs/$FUNCTION_NAME` by default.  The output
stream is configurable via the DNS_VM_GC_REPORTING_LOG_STREAM
environment variable.