Never convert properties to a timestamp. (#344)
This is a backwards incompatible change and will break loads with existing
schemas that use timestamp types for some properties. Sorry.

Previously we would attempt to recognize when a value is a date or timestamp and
use the proper BigQuery type. Unfortuantly these types are not declared in our
API schema and it's possible for the values to be null breaking type inference.
This then caused BigQuery loads to fail inconsistently. This fix removes the
type timestamp inference so we only use detect bool and numberic types which are
declared by GCP api discovery documents.

Other fixes with this checkin include:

- Upgrade to latest Beam SDK and associated libraires.
- Make it easier to use pip to install test dependencies.
- Describe how to supply num_shards paramerer in README.md (Issue #333).