Bigtable: how can we change the key of an existing table?
Designing the right key is crucial for optimal performance. But if we are not
happy with the results with our current key, how can we update it?

This example contains a Dataflow pipeline to read data from a table in a
Bigtable instance, and to write the same records to another table with the same
schema, but using a different key.

The example does not assume any specific schema and can work with any table.