Fix bugs with multiple disks and custom device names
Here are the current bugs with this script with respective fixes

key_name variable assignment is inside the for loop. It fails on the second iteration, as the key_name keeps getting added to its previous value. This is fixed by moving that line outside of the for loop, next to zone and region assignments
The API call for detach disk needs DeviceName field and not disk_name. Added the variable device_name to get the DeviceName and used it in the detach_disk function call
Line #294 has an error with "end" which is not expected there with only one %s for "operation".